% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{ECME}
\alias{ECME}
\title{The ECME function for the linear regression model based on the matrix-variant skewT distribution.}
\usage{
ECME(EM_type, N_cores, N_wait, Sigma_type, log_pdf_type, Y, X, ti)
}
\arguments{
\item{EM_type}{Must be one of "ADECME", "RPECME" and "RNECME".}

\item{N_cores}{The number of cores in the computer.0}

\item{N_wait}{Wait for the number of cores to complete the E steps. Must be less than num_cores.}

\item{Sigma_type}{A character. Must be either "DEC" or "diagonal".}

\item{log_pdf_type}{A character. Must be either "skewT" or "N".}

\item{Y}{The list of response matrices.}

\item{X}{The list of design matrices.}

\item{ti}{The list of measurement times.}
}
\value{
NULL.
}
\description{
The ECME function for the linear regression model based on the matrix-variant skewT distribution.
}
