% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MST_logpdf_c}
\alias{MST_logpdf_c}
\title{Rcpp function to calculate the log-likelihood for all observations}
\usage{
MST_logpdf_c(Y, M, A, Sigma, Psi, nu, log_pdf_type, Sigma_type)
}
\arguments{
\item{Y}{The list of response matrices.}

\item{M}{The list of location matrices.}

\item{A}{The list of skewness matrices.}

\item{Sigma}{The list of scale matrices.}

\item{Psi}{The column covariance matrix.}

\item{nu}{The degree of freedom}

\item{log_pdf_type}{Can be either "skewT" or "N".}

\item{Sigma_type}{Can be either "DEC" or "diagonal".}
}
\value{
The evaluated log-likelihood.
}
\description{
Rcpp function to calculate the log-likelihood for all observations
}
