% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD_ECM.R
\name{reg_simulation}
\alias{reg_simulation}
\title{Simulation program for the regression model}
\usage{
reg_simulation(
  N,
  beta_mat,
  A_vec,
  DEC_vec,
  Psi_mat,
  nu,
  Sigma_type,
  log_pdf_type
)
}
\arguments{
\item{N}{The number of subject.}

\item{beta_mat}{The predictor coefficients (q by p matrix). q is the number of predictors. p is the number of columns of Yi.}

\item{A_vec}{A numeric vector whose length is p.}

\item{DEC_vec}{A numeric vector whose length is 2.}

\item{Psi_mat}{A p by p matrix.}

\item{nu}{The degree of freedom.}

\item{Sigma_type}{The type of the correlation matrix. Must be "DEC" or "diagonal".}

\item{log_pdf_type}{Must be "skewT" or "N".}
}
\value{
The list of simulated data.
}
\description{
Simulation program for the regression model
}
\examples{
N <- 2500
beta_mat <- matrix(c(0.5,1.5,-0.5,0.5,1.5,-0.5),nrow = 3)
A_vec <- c(2.0,-2.0)
DEC_vec <- c(0.9,0.8)
Psi_mat <- matrix(c(1.0,-0.5,-0.5,1.0), nrow = 2)
nu <- 5.0
Sigma_type <- "DEC"
log_pdf_type <- "skewT"
simulated_data <- reg_simulation(N,
                                 beta_mat,
                                 A_vec,
                                 DEC_vec,
                                 Psi_mat,
                                 nu,
                                 Sigma_type,
                                 log_pdf_type)
print(str(simulated_data))
}
