% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMATREG-package.R
\name{STMATREG-package}
\alias{STMATREG-package}
\title{The 'STMATREG' package.}
\description{
This R package provides a linear regression based on the matrix variant skew-t distribution with the asynchronous distributed algorithm.
}
