% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{ECME}
\alias{ECME}
\title{The ECME function for the linear regression model based on the matrix-variant skewT distribution.}
\usage{
ECME(
  EM_type,
  N_cores,
  N_wait,
  Sigma_type,
  log_pdf_type,
  Y,
  X,
  ti,
  beta_value,
  A_value,
  DEC_value,
  Psi_value,
  nu_value
)
}
\arguments{
\item{EM_type}{must be one of "ADECME", "RPECME" and "RNECME".}

\item{N_cores}{the number of cores in the computer.0}

\item{N_wait}{wait for the number of cores to complete the E steps. Must be less than num_cores.}

\item{Sigma_type}{a character. Must be either "DEC" or "diagonal".}

\item{log_pdf_type}{a character. Must be either "skewT" or "N".}

\item{Y}{the list of response matrices.}

\item{X}{the list of design matrices.}

\item{ti}{the list of measurement times.}

\item{beta_value}{the initial value of beta.}

\item{A_value}{the initial value of A.}

\item{DEC_value}{the initial value of DEC.}

\item{Psi_value}{the initial value of Psi.}

\item{nu_value}{the initial value of nu.}
}
\value{
NULL.
}
\description{
The ECME function for the linear regression model based on the matrix-variant skewT distribution.
}
